refactor(core): break locatorUtil <-> error import cycle via a leaf helper module (#1051)
#1051 made locatorUtil import LocatorResolutionError, which imports
getLocatorInfoForErrorLog back from locatorUtil — a circular dependency
between the locator resolver and the error hierarchy that is fragile in CJS
bundles and hard to reason about.

Move the pure getLocatorInfoForErrorLog helper into its own leaf module
(imports only the locator model) and re-point all interactor error classes at
it, so the error hierarchy no longer imports locatorUtil. locatorUtil
re-exports the helper, so locatorUtil.getLocatorInfoForErrorLog stays on the
public surface (core.api.md unchanged). No behavior change.
