+++ packages/core/src/errors/WaitForFailureError.ts
@@ -1,6 +1,6 @@
 import { WaitForOption } from '../drivers/WaitForOption';
 import { PartLocator } from '../locators';
-import { getLocatorInfoForErrorLog } from '../utils/locatorUtil';
+import { getLocatorInfoForErrorLog } from '../utils/getLocatorInfoForErrorLog';
 import { InteractorErrorBase } from './InteractorErrorBase';
 
 export const WaitForFailureErrorId = 'WaitForFailureError';
