+++ packages/core/src/utils/locatorUtil.ts
@@ -180,20 +180,8 @@ export function overrideLocatorRelativePosition(
   });
 }
 
-/**
- * Display a rough description of the locators for error logging
- * this is an estimate, not a precise description with the absence of interactor
- * locators such as LinkedCssLocator would not be interpreted correctly
- * @param locator
- * @returns
- */
-export function getLocatorInfoForErrorLog(locator: PartLocator): string {
-  const locators = Array.isArray(locator) ? locator : [locator];
-
-  const selectors: string[] = [];
-  for (const loc of locators) {
-    selectors.push(loc.selector);
-  }
-
-  return selectors.join(', ');
-}
+// Re-exported from a leaf module so the interactor errors can build their
+// `locatorDescription` from it without importing `locatorUtil` (which throws
+// LocatorResolutionError) — breaking a `locatorUtil ↔ error` import cycle while
+// keeping `locatorUtil.getLocatorInfoForErrorLog` on the public surface.
+export { getLocatorInfoForErrorLog } from './getLocatorInfoForErrorLog';
